refactor: according to PR comms
