+++ .eslintrc.json
@@ -9,6 +9,7 @@
       {
         "endOfLine": "auto"
       }
-    ]
+    ],
+    "@next/next/no-img-element": "off"
   }
 }
