+++ src/pages/characters/index.tsx
@@ -1,17 +1,17 @@
 import { useState } from "react";
 import { useQuery } from "@apollo/client";
 import { graphql } from "../../gql/";
-import { NextRouter, useRouter } from "next/router";
+import { NextPageContext } from "next";
+import { useRouter } from "next/router";
 import Link from "next/link";
 import InfiniteScroll from "react-infinite-scroll-component";
 import { ColorRing } from "react-loader-spinner";
-import Image from "next/image";
 import CharactersCardList from "../../components/CardLists/CharactersCardList";
 import SearchInput from "../../components/SearchInput";
 import Select from "../../components/Select";
 import useDebounce from "../../hooks/useDebounce";
-import logo from "../../../public/logo.png";
-import { NextPageContext } from "next";
+import getServerSideQueryParamFromContext from "../../helpers/getServerSideQueryParamFromContext";
+import updateQueryParamForRouter from "../../helpers/routerQueryManipulation";
 
 const genderOptions = ["Female", "Male", "Genderless", "Unknown"];
 const lifeStatusOptions = ["Alive", "Dead", "Unknown"];
@@ -37,63 +37,6 @@ const GET_CHARACTERS_QUERY = graphql(/* GraphQL */ `
     }
   }
 `);
-
-const getServerSideQueryParamFromContext = (context: NextPageContext, queryParamKey: string) => {
-  return context.query[queryParamKey] || null;
-};
-
-const removeQueryParamsFromRouter = ({ router, queryParamKey }: { router: NextRouter; queryParamKey: string }) => {
-  delete router.query[queryParamKey];
-
-  return router.replace(
-    {
-      pathname: router.pathname,
-      query: router.query,
-    },
-    undefined,
-    { shallow: true }
-  );
-};
-
-const addQueryParamToRouter = ({
-  router,
-  queryParamKey,
-  queryParamValue,
-}: {
-  router: NextRouter;
-  queryParamKey: string;
-  queryParamValue?: string;
-}) => {
-  return router.push(
-    {
-      pathname: router.pathname,
-      query: {
-        ...router.query,
-        [queryParamKey]: queryParamValue,
-      },
-    },
-    undefined,
-    {
-      shallow: true,
-    }
-  );
-};
-
-const updateQueryParamForRouter = ({
-  router,
-  queryParamKey,
-  queryParamValue,
-}: {
-  router: NextRouter;
-  queryParamKey: string;
-  queryParamValue?: string;
-}) => {
-  if (queryParamValue === "") {
-    return removeQueryParamsFromRouter({ router, queryParamKey });
-  } else {
-    return addQueryParamToRouter({ router, queryParamKey, queryParamValue });
-  }
-};
 
 function useCharactersList({ name, gender, status }: CharactersPagePageProps) {
   const router = useRouter();
@@ -188,9 +131,9 @@ const CharactersPage = ({ name, gender, status }: CharactersPagePageProps) => {
   return (
     <div className='w-full min-h-full bg-black/[.85] text-[14px] [background-image:url("../../public/endless-constellation.svg")]'>
       <div className='flex flex-col md:flex-row md:justify-center lg:justify-start items-center'>
-        <Image
+        <img
           className='w-[200px] h-[200px] md:w-[240px] md:h-[240px] xl:w-[300px] xl:h-[300px] max-md:ml-0 ml-8 pointer-events-none select-none'
-          src={logo}
+          src={"logo.png"}
           alt={"Rick'n'Morty Logo"}
           draggable={false}
         />
@@ -249,8 +192,7 @@ const CharactersPage = ({ name, gender, status }: CharactersPagePageProps) => {
           next={handleFetchMore}
           hasMore={hasMore}
           loader={
-            <h4 className='w-full flex justify-center font-mali'>
-              Loading more
+            <div className='w-full flex justify-center font-mali'>
               <ColorRing
                 visible={true}
                 height='80'
@@ -258,7 +200,7 @@ const CharactersPage = ({ name, gender, status }: CharactersPagePageProps) => {
                 ariaLabel='blocks-loading'
                 colors={["#bfd84d", "#9db33c", "#429EA6", "#12b0c9", "#51E5FF"]}
               />
-            </h4>
+            </div>
           }
           endMessage={
             dataLength === 0 ? (
